Separate the charts into modules
