+++ AlertaDengue/dados/charts/home.py
@@ -1,268 +1,181 @@
 from datetime import timedelta
 from time import mktime
-import json
 from django.utils.translation import gettext as _
 
-import dash_core_components as dcc
-import dash_html_components as html
-from dash.dependencies import Input, Output
 import plotly.graph_objs as go
-from django_plotly_dash import DjangoDash
-
-import plotly.graph_objects as go
 import pandas as pd
 from plotly.subplots import make_subplots
-import numpy as np
-import dash_core_components as dcc
-import plotly.express as px
-
 
 # local
-from .dbdata import get_series_by_UF, load_series, sql_home_charts
+from dados.dbdata import get_series_by_UF
+
+'''
+Module for plotting in the homepage charts
+'''
 
 
-def home_chart1():
-    animals = ['giraffes', 'orangutans', 'monkeys']
-
-    fig = go.Figure(
-        data=[
-            go.Bar(name='SF Zoo', x=animals, y=[20, 14, 23]),
-            go.Bar(name='LA Zoo', x=animals, y=[12, 18, 29]),
-        ]
-    )
-    # Change the bar mode
-    fig.update_layout(
-        barmode='group',
-        autosize=False,
-        width=295,
-        height=145,
-        margin=dict(l=50, r=50, b=10, t=10, pad=4),
-        paper_bgcolor="LightSteelBlue",
-        legend=dict(
-            x=0,
-            y=1.0,
-            bgcolor='rgba(255, 255, 255, 0)',
-            bordercolor='rgba(255, 255, 255, 0)',
-        ),
-    )
-
-    graph = fig.to_html(full_html=False, default_height=270, default_width=350)
-    return graph
-
-
-def home_chart2():
-    # Stack chart
-    animals = ['giraffes', 'orangutans', 'monkeys']
-
-    fig2 = go.Figure(
-        data=[
-            go.Bar(name='SF Zoo', x=animals, y=[20, 14, 23]),
-            go.Bar(name='LA Zoo', x=animals, y=[12, 18, 29]),
-        ]
-    )
-    # Change the bar mode
-    fig2.update_layout(
-        barmode='stack',
-        autosize=False,
-        width=275,
-        height=145,
-        margin=dict(l=50, r=50, b=10, t=10, pad=4),
-        paper_bgcolor="LightSteelBlue",
-        legend=dict(
-            x=0,
-            y=1.0,
-            bgcolor='rgba(255, 255, 255, 0)',
-            bordercolor='rgba(255, 255, 255, 0)',
-        ),
-    )
-
-    graph2 = fig2.to_html(
-        full_html=False, default_height=270, default_width=350
-    )
-    return graph2
-
-
-def home_chart3():
-    # Chart line
-    x = np.arange(10)
-
-    fig3 = go.Figure(data=go.Scatter(x=x, y=x ** 2))
-    fig3.update_layout(
-        barmode='stack',
-        autosize=False,
-        width=275,
-        height=145,
-        margin=dict(l=50, r=50, b=10, t=10, pad=4),
-        paper_bgcolor="LightSteelBlue",
-        legend=dict(
-            x=0,
-            y=1.0,
-            bgcolor='rgba(255, 255, 255, 0)',
-            bordercolor='rgba(255, 255, 255, 0)',
-        ),
-    )
-
-    graph3 = fig3.to_html(
-        full_html=False, default_height=270, default_width=350
-    )
-    return graph3
-
-
-def home_chart4():
-    # Scatter
-    N = 1000
-    t = np.linspace(0, 10, 100)
-    y = np.sin(t)
-
-    fig4 = go.Figure(data=go.Scatter(x=t, y=y, mode='markers'))
-    fig4.update_layout(
-        barmode='stack',
-        autosize=False,
-        width=275,
-        height=145,
-        margin=dict(l=50, r=50, b=10, t=10, pad=4),
-        paper_bgcolor="LightSteelBlue",
-        legend=dict(
-            x=0,
-            y=1.0,
-            bgcolor='rgba(255, 255, 255, 0)',
-            bordercolor='rgba(255, 255, 255, 0)',
-        ),
-    )
-
-    graph4 = fig4.to_html(
-        full_html=False, default_height=270, default_width=350
-    )
-    return graph4
-
-
-external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
-
-app = DjangoDash('SimpleExample', external_stylesheets=external_stylesheets)
-
-app.layout = html.Div(
-    [
-        html.H1('Square Root Slider Graph'),
-        dcc.Graph(
-            id='slider-graph',
-            animate=True,
-            style={"backgroundColor": "#1a2d46", 'color': '#ffffff'},
-        ),
-        dcc.Slider(
-            id='slider-updatemode',
-            marks={i: '{}'.format(i) for i in range(20)},
-            max=20,
-            value=2,
-            step=1,
-            updatemode='drag',
-        ),
-    ]
-)
-
-
-@app.callback(
-    Output('slider-graph', 'figure'), [Input('slider-updatemode', 'value')]
-)
-def display_value(value):
-
-    x = []
-    for i in range(value):
-        x.append(i)
-
-    y = []
-    for i in range(value):
-        y.append(i * i)
-
-    graph = go.Scatter(x=x, y=y, name='Manipulate Graph')
-    layout = go.Layout(
-        paper_bgcolor='#27293d',
-        plot_bgcolor='rgba(0,0,0,0)',
-        xaxis=dict(range=[min(x), max(x)]),
-        yaxis=dict(range=[min(y), max(y)]),
-        font=dict(color='white'),
-    )
-    return {'data': [graph], 'layout': layout}
-
-
-def home_chart_bar():
-    df = sql_home_charts()
-
-    df_home = df[['SE', 'casos_est', 'casos', 'nome', 'uf', 'Rt']]
-
-    filter_se = df_home.SE > 202015
-
-    select_uf = df_home['uf'].map(lambda x: x.startswith('Ceará'))
-
-    df_uf = df_home[select_uf][filter_se].sort_values('uf')
-    df_ufs = df_uf.groupby('SE').sum()
-
-    # import pdb; pdb.set_trace()
+class HomeCharts:
     colors = {
-        'A': px.colors.qualitative.Light24[2],
-        'B': px.colors.qualitative.Pastel2[1],
+        'Ceará': 'rgb(0,0,0)',
+        'Espírito Santo': 'rgb(255,0,0)',
+        'Paraná': 'rgb(0,255,0)',
+        'Minas Gerais': 'rgb(0,0,255)',
+        'Rio de Janeiro': 'rgb(255,255,0)',
+        'São Paulo': 'rgb(0,255,255)',
+        'Rio Grande do Sul': 'rgb(255,150,0)',
+        'Maranhão': 'rgb(255,0,255)',
     }
 
-    fig = go.Figure(
-        data=[
-            go.Bar(
-                name='Casos estatísticos',
-                x=df_ufs.index.map(
-                    lambda v: '%s/%s' % (str(v)[:4], str(v)[-2:])
-                ),
-                y=df_ufs.casos_est,
-                text=df_ufs.index.map(lambda v: '{}'.format(str(v)[-2:])),
-                hoverinfo='text',
-                hovertemplate="Semana %{text}<br>Incidência=%{y:1f}",
-                marker={'color': colors['B']}
-                #         marker_color=
-            ),
-            go.Scatter(
-                name='Casos',
-                x=df_ufs.index.map(
-                    lambda v: '%s/%s' % (str(v)[:4], str(v)[-2:])
-                ),
-                y=df_ufs.casos,
-                text=df_ufs.index.map(lambda v: '{}'.format(str(v)[-2:])),
-                hoverinfo='text',
-                hovertemplate="Semana %{text}<br>Incidência=%{y:1f}",
-                marker={'color': colors['A']}
-                #         marker_color=
-            ),
-        ]
-    )
-    # Change the bar mode
-    fig.update_layout(
-        xaxis=dict(
-            title='Período (Ano/Semana)',
-            tickangle=-60,
-            nticks=4,
-            showline=False,
-            showgrid=True,
-            showticklabels=True,
-            linecolor='rgb(204, 204, 204)',
-            linewidth=0,
-            gridcolor='rgb(176, 196, 222)',
-            ticks='outside',
-            tickfont=dict(family='Arial', size=12, color='rgb(82, 82, 82)'),
-        ),
-        yaxis=dict(
-            title='Incidência',
-            showline=False,
-            showgrid=True,
-            showticklabels=True,
-            linecolor='rgb(204, 204, 204)',
-            linewidth=0,
-            gridcolor='rgb(176, 196, 222)',
-        ),
-        showlegend=True,
-        plot_bgcolor='rgb(255, 255, 255)',
-        paper_bgcolor='rgb(245, 246, 249)',
-        autosize=False,
-        width=575,
-        height=245,
-        margin=dict(l=50, r=50, b=10, t=30, pad=4),
-    )
-    fig.update_layout(title=df_uf.uf.iloc[0])
+    @classmethod
+    def total_series(cls, case_series, disease):
+        '''
+        :param case_series:
+        :param disease: dengue|chikungunya|zika
+        :return:
+        '''
+        # gc = context['geocodigos'][0]
+        series = (
+            get_series_by_UF(disease, weeks=52)
+            if disease not in case_series
+            else case_series[disease]
+        )
 
-    graph = fig.to_html(full_html=False, default_height=270, default_width=350)
-    return graph
+        if series.empty:
+            return {
+                'ufs': [],
+                'start': None,
+                'series': {},
+                'series_est': {},
+                'disease': disease,
+            }
+
+        ufs = list(set(series.uf.tolist()))
+        # 51 weeks to get the end of the SE
+        start = series.data.max() - timedelta(weeks=51)
+        start = int(mktime(start.timetuple()))
+        casos = {}
+        casos_est = {}
+        initial_range = -52  # to get the last 52 weeks
+
+        for uf in ufs:
+            series_uf = series[series.uf == uf]
+            datas = [
+                int(mktime(d.timetuple())) * 1000
+                for d in series_uf.data[initial_range:]
+            ]
+            casos[uf] = [
+                list(t)
+                for t in zip(datas, series_uf.casos_s[initial_range:].tolist())
+            ]
+            casos_est[uf] = [
+                list(t)
+                for t in zip(
+                    datas, series_uf.casos_est_s[initial_range:].tolist()
+                )
+            ]
+
+        return {
+            'ufs': ufs,
+            'start': start,
+            'series': casos,
+            'series_est': casos_est,
+            'disease': disease,
+        }
+
+    @classmethod
+    def _create_chart(cls, case_series, disease):
+        series_est = cls.total_series(case_series, disease=disease)[
+            'series_est'
+        ]
+
+        dfs = []
+        for k, v in series_est.items():
+            df = pd.DataFrame(v)
+            df.set_index(pd.to_datetime(df[0], unit='ms'), inplace=True)
+            df.drop(columns=0, inplace=True)
+            df.rename(columns={1: k}, inplace=True)
+            df.index.name = None
+            dfs.append(df)
+
+        df_ufs = pd.concat(dfs, sort=True)
+
+        fig = make_subplots(specs=[[{"secondary_y": True}]])
+
+        for k in df_ufs:
+            fig.add_trace(
+                go.Scatter(
+                    x=df_ufs.index,
+                    y=df_ufs[k],
+                    name=k,
+                    marker={'color': cls.colors[k]},
+                    text=df_ufs.index.strftime('%d-%b-%Y <br>{}'.format(k)),
+                    hovertemplate=_(
+                        '%{text} <br>' '%{y} Casos Estimados' '<extra></extra>'
+                    ),
+                ),
+                secondary_y=True,
+            )
+
+        fig.update_layout(
+            height=350,
+            width=1000,
+            title=go.layout.Title(
+                text=_(
+                    'Casos Estimados de {} ' 'nos municípios monitorados'
+                ).format(disease.capitalize()),
+                font=dict(family="sans-serif", size=16),
+            ),
+            plot_bgcolor='rgb(255, 255, 255)',
+            paper_bgcolor='rgb(255, 255, 255)',
+            showlegend=True,
+            font=dict(family="sans-serif", size=12),
+            xaxis=dict(
+                # title='',
+                tickangle=-20,
+                nticks=len(df) // 3,
+                showline=True,
+                showgrid=True,
+                showticklabels=True,
+                linecolor='rgb(204, 204, 204)',
+                linewidth=0,
+                gridcolor='rgb(176, 196, 222)',
+                ticks='outside',
+                tickfont=dict(
+                    family='Arial', size=12, color='rgb(82, 82, 82)'
+                ),
+            ),
+            yaxis=dict(
+                # title='',
+                showline=True,
+                showgrid=True,
+                showticklabels=True,
+                linecolor='rgb(204, 204, 204)',
+                linewidth=0,
+                gridcolor='rgb(176, 196, 222)',
+            ),
+        )
+
+        fig.update_yaxes(
+            title_text=_('Casos'),
+            secondary_y=True,
+            showline=False,
+            showgrid=True,
+            showticklabels=True,
+            linecolor='rgb(204, 204, 204)',
+            linewidth=0,
+            gridcolor='rgb(204, 204, 204)',
+        )
+
+        return fig.to_html()
+
+    @classmethod
+    def create_dengue_chart(cls, case_series):
+        return cls._create_chart(case_series, 'dengue')
+
+    @classmethod
+    def create_chik_chart(cls, case_series):
+        return cls._create_chart(case_series, 'chikungunya')
+
+    @classmethod
+    def create_zika_chart(cls, case_series):
+        return cls._create_chart(case_series, 'zika')
